chore: rename and move utils to libs/global-analytics-helper
